Revise Migration and Seeder

+++ app/Database/Migrations/2023-08-24-060938_CreateBrandsTable.php
@@ -12,7 +12,7 @@ class CreateBrandsTable extends Migration
     {
         $this->forge->addField([
             'id' => [
-                'type' => 'bigint',
+                'type' => 'int',
                 'unsigned' => true,
                 'auto_increment' => true
             ],
